+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add `Resolution#records` method to query multiple records #96
 * Add formatting options for `Resolution#namehash` and `Resolution#childhash` #91
+* Plug-in network config from dot-crypto library #101
 
 ## 1.8.3